requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
dominate==2.9.1
email_validator==2.1.1
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.5
Flask-WTF==1.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
python-dotenv==1.0.1
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.2
